requirements.txt
redis
asyncio-nats-streaming
numpy